+++ src/main/java/tis/project/lion/postproject/api/controller/ApiError.java
@@ -1,4 +1,4 @@
-package tis.project.lion.postproject.domain;
+package tis.project.lion.postproject.api.controller;
 
 import org.springframework.http.HttpStatus;
