+++ src/main/java/tis/project/lion/postproject/api/controller/ApiResult.java
@@ -1,4 +1,4 @@
-package tis.project.lion.postproject.domain;
+package tis.project.lion.postproject.api.controller;
 
 import org.springframework.http.HttpStatus;
