+++ src/main/java/tis/project/lion/postproject/api/controller/post/PostDto.java
@@ -1,5 +1,7 @@
-package tis.project.lion.postproject.domain.post;
+package tis.project.lion.postproject.api.controller.post;
 
+
+import tis.project.lion.postproject.domain.post.Post;
 
 public class PostDto {
