+++ src/main/resources/application.yml
@@ -7,11 +7,6 @@ spring:
   jpa:
     hibernate:
       ddl-auto: create
-#  sql:
-#    init:
-#      platform: mysql
-#      schema-locations: classpath:schema-h2.sql
-#      data-locations: classpath:data.sql
 logging:
   level:
     root: info
