--- src/main/resources/data.sql
@@ -1,9 +1,0 @@
-insert into board(board_id, name) values (null, 'testBoard1');
-insert into board(board_id, name) values (null, 'testBoard2');
-
-insert into post(post_id, content, board_id, password, title, writer) VALUES (null, 'test_content', 1, 1234, 'title', 'tis');
-insert into post(post_id, content, board_id, password, title, writer) VALUES (null, 'test_content', 1, 1234, 'title', 'tis');
-insert into post(post_id, content, board_id, password, title, writer) VALUES (null, 'test_content', 1, 1234, 'title', 'tis');
-insert into post(post_id, content, board_id, password, title, writer) VALUES (null, 'test_content', 1, 1234, 'title', 'tis');
-insert into post(post_id, content, board_id, password, title, writer) VALUES (null, 'test_content', 1, 1234, 'title', 'tis');
-insert into post(post_id, content, board_id, password, title, writer) VALUES (null, 'test_content', 1, 1234, 'title', 'tis');
